apps/api/pyproject.toml
[project]
name = "portal-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "arq>=0.26",
    "pydantic>=2.9",
    "email-validator>=2.2",
    "pydantic-settings>=2.6",
    "sqlmodel>=0.0.22",
    "alembic>=1.14",
    "redis>=5.2",
    "sentry-sdk[fastapi]>=2.20",
    "bcrypt>=4.0",
    "pyjwt>=2.10",
    "python-multipart>=0.0.20",
    "httpx>=0.28",
    # OTel contrib/instrumentation packages have no stable 1.x yet; 0.5xb pins are intentional.
    "opentelemetry-distro>=0.50b0",
    "opentelemetry-exporter-otlp-proto-http>=1.29",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-instrumentation-logging>=0.50b0",
    "opentelemetry-instrumentation-redis>=0.50b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.50b0",
    "pillow>=10.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "fakeredis>=2.26",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "RUF"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
